+++ libursa/MemMap.h
@@ -1,8 +1,5 @@
 #pragma once
 
-#include <unistd.h>
-
-#include <cstddef>
 #include <cstdint>
 #include <stdexcept>
 #include <string>
