+++ libursa/OnDiskIndex.cpp
@@ -7,6 +7,7 @@
 
 #include "FeatureFlags.h"
 #include "Query.h"
+#include "QueryGraph.h"
 #include "Utils.h"
 #include "spdlog/spdlog.h"
