+++ libursa/QueryParser.h
@@ -1,6 +1,5 @@
 #pragma once
 
 #include "Command.h"
-#include "FeatureFlags.h"
 
 Command parse_command(const std::string &s);
